backend/requirements.txt
# Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
python-multipart==0.0.6

# Database
pymongo==4.6.0
motor==3.3.2
redis==5.0.1

# ML & Analytics
scikit-learn==1.3.2
pandas==2.1.3
numpy==1.26.2
nltk==3.8.1

# NLP & Text Processing
python-docx==1.1.0
PyPDF2==3.0.1
textract==1.6.5

# Data Validation
pydantic-settings==2.1.0
email-validator==2.1.0

# Monitoring & Logging
prometheus-client==0.19.0
python-json-logger==2.0.7

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.2

# Utils
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0
